+++ karma.conf.js
@@ -12,9 +12,6 @@ module.exports = function(config) {
             require('karma-coverage'),
             require('@angular-devkit/build-angular/plugins/karma')
         ],
-        files: [
-          'src/app/shared/mocks/google-api.js'
-        ],
         client: {
             jasmine: {},
             clearContext: false // leave Jasmine Spec Runner output visible in browser