+++ set-env.ts
@@ -50,7 +50,7 @@ function getEnvironment(featuress) {
   production: ${isProduction},
   environment: '${(isProduction) ? 'prod' : 'dev'}',
   domain: '${(isProduction) ? process.env.DOMAIN_PROD : process.env.DOMAIN_DEV}',
-  googleMapsApiKey: '${process.env.GOOGLE_MAPS_API_KEY}',
+  stadiaApiKey: '${process.env.STADIA_API_KEY || ''}',
   app: {
   modules: {
   mapMenu: ${featuress.mapMenu},