+++ src/app/main/map/components/map-menu/components/map-filter/map-filter.component.ts
@@ -19,10 +19,8 @@ export class MapFilterComponent implements OnInit {
   readonly faFilter = faFilter;
   readonly markerCut = Const.markerColor.cut;
   readonly markerRecovred = Const.markerColor.recovred;
-  readonly markerCutUser = Const.markerColor.cutUser;
   readonly reportSatus = {
     cut: ReportSatus.CUT,
-    cut_owner: ReportSatus.CUT_OWNER,
     recovred: ReportSatus.CUT_COMPLETED
   };
   clickedStatus: ReportSatus[] = [];