+++ src/environments/const.ts
@@ -38,13 +38,8 @@ export const Const = {
     support: 'support@bogal.ca'
   },
   markerColor: {
-    user: 'https://maps.google.com/mapfiles/ms/icons/blue-dot.png',
-    cut: 'https://maps.google.com/mapfiles/ms/icons/red-dot.png',
-    cutUser: 'https://maps.google.com/mapfiles/ms/icons/orange-dot.png',
-    recovred: 'https://maps.google.com/mapfiles/ms/icons/green-dot.png'
-    // user: '../assets/static/images/markers/marker_blue.svg',
-    // cut: '../assets/static/images/markers/marker_red.svg',
-    // cutUser: '../assets/static/images/markers/marker_orange.svg',
-    // recovred: '../assets/static/images/markers/marker_green.svg'
+    user: 'assets/static/images/markers/marker_blue.svg',
+    cut: 'assets/static/images/markers/marker_red.svg',
+    recovred: 'assets/static/images/markers/marker_green.svg'
   }
 };